### 2024-07-30 #7053 @editorialbot
[PRE REVIEW]: colorspace: A Python Toolbox for Manipulating and Assessing Colors and Palettes
**Submitting author:** @retostauffer (<a href="http://orcid.org/0000-0002-3798-5507">Reto Stauffer</a>)
**Repository:** https://github.com/retostauffer/python-colorspace
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.0
