### 2024-07-30 #7053 @danielskatz
@retostauffer - thanks for your submission.  As you can see, our DOI checked failed on your bibtex file. I think it might be because of the commented-out entries, as I don't think the use of `%` is really standard.  Can you perhaps remove the parts of the bibtex file that are not needed here?  And then use the command `@editorialbot check references` to check the file again. editorialbot commands need to be the first entry in a new comment.
