### 2024-07-30 #7053 @danielskatz
Additionally, it's not clear to me that this is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS, so the editors will now discuss this. You should hear back in a week or two.
