### 2024-07-30 01d51d8
fix: Removing commented entries from bibtex file
