+++ paper_assets/paper.bib
@@ -143,8 +143,6 @@
   publisher = {Springer Science and Business Media {LLC}}
 }
 
-
-# Software packages providing color maps
 @Manual{colormap,
   title     = {Colormap},
   author    = {Cokelaer, Thomas},
@@ -189,77 +187,3 @@
   url       = {https://pypi.org/project/cmcrameri/}
 }
 
-
-%@InCollection{Baum:2019,
-%  title = {An Evaluation of Color Maps for Visual Data Exploration},
-%  author = {Baum, Daniel},
-%  year = {2019},
-%  publisher = {De Gruyter},
-%  booktitle = {Science in Color},
-%  isbn = {9783110605211},
-%  doi = {10.1515/9783110605211-010},
-%  pages = {147--162}
-%}
-
-%@Book{Wilkinson:2005,
-%  author = {Wilkinson, Leland},
-%  title = {The Grammar of Graphics},
-%  year = {2005},
-%  edition = {2nd},
-%  publisher = {Springer-Verlag},
-%  addres = {New York}
-%}
-
-%@Manual{plotly,
-%  author = {{Plotly Technologies Inc.}},
-%  title = {Collaborative Data Science},
-%  address = {Montreal, QC},
-%  year = {2015},
-%  url = {https://plot.ly},
-%  note = {accessed 2024-07-24}
-%}
-
-%@Book{Kaiser:1996,
-%  author = {Kaiser, Peter K. and Boynton, Robert M.},
-%  title = {Human Color Vision},
-%  publisher = {Optical Society of America},
-%  year = {1996},
-%  edition = {2nd},
-%  address = {Washington, DC}
-%}
-
-%@InCollection{Knoblauch:2002,
-%  author = {Kenneth Knoblauch},
-%  title = {Color Vision},
-%  booktitle = {Steven's Handbook of Experimental Psychology -- Sensation and Perception},
-%  editor = {Yantis, Steven and Pashler, Hal},
-%  year = {2002},
-%  edition = {3rd},
-%  volume = {1},
-%  publisher = {John Wiley \& Sons},
-%  address = {New York},
-%  pages = {41--75}
-%}
-
-%@Article{Vienot:1995,
-%  author = {Vi{\'e}not, F. and Brettel, H. and Ott, L. and M'Barek, A. Ben and Mollon,  J. D.},
-%  title = {What Do Colour-Blind People See?},
-%  journal = {Nature},
-%  year = {1995},
-%  day = {1},
-%  volume = {376},
-%  number = {6536},
-%  pages = {127--128},
-%  doi = {10.1038/376127a0}
-%}
-
-%@Article{Zeileis:2007,
-%  author = {Zeileis, Achim and Meyer, David and Hornik, Kurt},
-%  title = {Residual-Based Shadings for Visualizing (Conditional) Independence},
-%  journal = {Journal of Computational and Graphical Statistics},
-%  year = {2007},
-%  volume = {16},
-%  number = {3},
-%  pages = {507--525},
-%  doi = {10.1198/106186007x237856},
-%}
