### 2024-07-30 #7053 @retostauffer
Removed commented-out lines in the bibtex file. Thank you for the hint, @danielskatz .
