### 2024-07-30 #7053 @zeileis
The DOI [10.1016/0306-4573(91)90096-5](https://doi.org/10.1016/0306-4573(91)90096-5) is for a review of the Tufte (1990) book _Envisioning Information_ and not the book itself. For the book itself we couldn't find a DOI and hence did not further modify the .bib file.
