### 2024-08-08 #7053 @danielskatz
@retostauffer - thanks for waiting. We've decided this is research software, though it wasn't completely clear. I hope that during the review, you can make the case more clearly in the paper and in the readme/documentation, such as what kinds of research this is useful for.
