### 2024-08-09 #7053 @retostauffer
We'd like to thank @danielskatz for handling the initial submission to JOSS and accepting our article for entering review. The choice of effective color maps plays a crucial role in accurately communicating scientific results, which are often presented visually through graphs and figures. In the revised version of the manuscript, we will more clearly emphasize the significance of this aspect for science and research.

Additionally, we would like to extend our thanks to @jromanowska for overseeing the review process. Unfortunately, we do not have any suitable reviewers to suggest at this time, but I will reach out if I am able to identify any appropriate candidates.
