### 2024-08-09 #7053 @danielskatz
@retostauffer - as is written in the first comment in this issue, you might want to search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
