### 2024-08-19 #7053 @jromanowska
:wave: @hollowscene , thank you for volunteering to review this submission! Please, feel free to look around and check [the docs](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), while I try to find reviewer no.2.
