### 2024-08-20 #7120 @editorialbot
[REVIEW]: colorspace: A Python Toolbox for Manipulating and Assessing Colors and Palettes
**Submitting author:** @retostauffer (<a href="http://orcid.org/0000-0002-3798-5507">Reto Stauffer</a>)
**Repository:** https://github.com/retostauffer/python-colorspace
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
