### 2024-08-20 #7120 @jromanowska
```@editorialbot generate my checklist```

I encourage you to **over-communicate** and let everyone know that you're on the task every now and then, instead of waiting several weeks to collect all your comments, questions, or suggestions.

_Please feel free to ping me (@jromanowska) if you have any questions/concerns._
