### 2024-08-20 #7120 @hollowscene
Hey everyone, just thought I'd make an early start to some of the housekeeping tickboxes. As a heads up - I'll most likely hold off on actually installing and playing around with the package until I've got a bit more time this weekend.

I hate to be nitpicky but just for the statement of need tickboxes, I feel like I'm not getting an explicit designation of who the target audience is. I definitely acknowledge it does feel a bit redundant for this package considering how ubiquitous its usage should be, and it sort of is implied within the current statement of need. But just wanted to double check this before ticking it off as I'm not entirely sure how strict we need to be!
