### 2024-08-20 #7120 @jromanowska
Good point! The tickboxes are very general - to suit all the different types of software submitted to JOSS. As long as you feel that the "Statement of need" section in the paper is alright, you can check off the box. However, you could also ask the authors to have a second look at it and improve if you feel something is missing.

And thanks for starting right off - we encourage here over-commenting, so let's keep it like that :+1:
