### 2024-09-04 #7120 @dmreagan
I also wondered about the explicit target audience as in https://github.com/openjournals/joss-reviews/issues/7120#issuecomment-2298631975. But I think it is implied that the intended users are the same users of the "most graphics packages" the authors refer to in the second paragraph. It's a ✅ for me.
