### 2024-09-04 #7120 @dmreagan
Question about references: is it OK to have web links in the body of the paper? The authors have included direct URLs to a few of their own web resources while putting all 3rd-party links down in the References section.
