### 2024-09-05 #7120 @jromanowska
> Question about references: is it OK to have web links in the body of the paper? The authors have included direct URLs to a few of their own web resources while putting all 3rd-party links down in the References section.

For me, that seems logical. However, if you think it doesn't suit, go ahead and suggest another way of doing it. JOSS does not have strict formatting rules for the papers.
