### 2024-09-07 retostauffer/python-colorspace#21 @hollowscene
Custom palette tests fail on Windows
Five tests fail in `test_palettes_custom_palconfig_files.py` when test suite is run on Windows.

Here is the full pytest output: [pytest_results.txt](https://github.com/user-attachments/files/16918444/pytest_results.txt)

For quick access, the error message of one of those test fails:

```
___________________________________ test_hclpalettes_custom_palconfig_good_custom_div ___________________________________

    def test_hclpalettes_custom_palconfig_good_custom_div():

        tmpfile = NamedTemporaryFile()
>       with open(tmpfile.name, "w") as fid: fid.write(good_custom_div)
E       PermissionError: [Errno 13] Permission denied: 'C:\\Users\\andrew\\AppData\\Local\\Temp\\tmpmchuptx8'
```

This appears to be an OS difference caused by the tempfile package. [Old but relevant StackOverflow answer](https://stackoverflow.com/a/3924253).
