### 2024-09-07 6cdde0d
fix: update custom palettes test to run with windows os
