### 2024-09-08 #7120 @hollowscene
Just finished my review checklist.

I did encounter one issue with some tests failing due to some Windows OS discrepancies which I've raised a GitHub issue for. Regardless, automated tests are thorough and well documented so I've still ticked it off in the checklist.

Regarding references, adding 3rd-party links seems reasonable and is more quickly accessible so I've also given it a tick.

Overall, the paper is well written and the package has had significant work dedicated to it. I have no doubt this package will be incredibly useful for both casual and veteran data vizzers. Amazing work :)
