### 2024-09-08 #7120 @zeileis
Andrew @hollowscene, thank you so much for the positive and constructive feedback along with the issues and pull requests in the GitHub project. Reto is currently on vacation but we'll incorporate all your helpful suggestions in the software and paper when he's back! Thanks & best wishes, Achim
