### 2024-09-10 retostauffer/python-colorspace#23 @dmreagan
setuptools not always included in virtualenv
I'm testing the installation procedures for the review at openjournals/joss-reviews#7120. I tried following the [instructions](https://retostauffer.github.io/python-colorspace/installation.html#installation) for installing from a git clone.

```
git clone https://github.com/retostauffer/python-colorspace.git
cd python-colorspace && python setup.py install
```

It worked just fine from a conda environment, but failed when I tried a virtualenv environment because `setuptools` was missing. I am on a Ubuntu system. It might be worth adding a line to the documentation that not all virtual environments will include `setuptools` by default, or else find some way to ensure that `setuptools` is there.
