### 2024-09-11 retostauffer/python-colorspace#24 @dmreagan
Error message when running tests with the makefile
I'm testing the installation procedures for the review at https://github.com/openjournals/joss-reviews/issues/7120. I tried following the [instructions](https://retostauffer.github.io/python-colorspace/installation.html#installation) for running the tests with the makefile.

```
# Setting up virtual environment with all dependencies for full
# tests as well as for creating the documentation, calculating
# coverage etc., install Python colorspace, activate venv and
# run full tests.
make venv && make install && source venv/bin/activate && make test
```

I tried this on two Ubuntu systems running the bash shell and both times got the following error:

```
source venv/bin/activate && make install
/bin/sh: 1: source: not found
make: *** [Makefile:31: venv] Error 127
```
