### 2024-09-11 retostauffer/python-colorspace#24 @zeileis
David, thanks for the report, very much appreciated! I can replicate this on my end (Debian/testing) and I think that the culprit is the last line of the `venv` target in the `Makefile` (L31).

To confirm, could you try to run the following after you got the error above?

```
source venv/bin/activate
make install
make test
```

For me this works and successfully completes all tests.
