### 2024-09-12 retostauffer/python-colorspace#24 @dmreagan
Yes, those commands did work for me and the tests passed.
