### 2024-09-12 retostauffer/python-colorspace#24 @zeileis
Thanks, David, I've created PR #25 that tries to resolve the problem. But I'll let Reto have a final look when he returns from his vacations.
