### 2024-10-01 2d8733b
fix (#21): Adjusting handling of named temporary files to be complient with how Windows
