+++ src/colorspace/tests/test_palettes_custom_palconfig_files.py
@@ -1,5 +1,13 @@
+# ---------------------------------------------------
+# Note: To be complient with Windows file access
+# NamedTemporaryFiles are handled as follows:
+#
+# 1. Create named temporary file with delete = False
+# 2. Write content, flush and close file connection
+# 3. At the end, delete temporary file via os.remove
+# ---------------------------------------------------
 
-
+import os
 import pytest
 from pytest import raises
 from colorspace import hclpalettes
@@ -77,8 +85,11 @@ def test_hclpalettes_custom_palconfig_wrong_usage():
 # ------------------------------------------
 def test_hclpalettes_custom_palconfig_good_custom_div():
 
-    tmpfile = NamedTemporaryFile()
-    with open(tmpfile.name, "w") as fid: fid.write(good_custom_div)
+    tmpfile = NamedTemporaryFile(delete = False)
+    with open(tmpfile.name, "w") as fid:
+        fid.write(good_custom_div)
+        fid.flush()
+        fid.close()
 
     pals1 = hclpalettes(files = tmpfile.name)
     pals2 = hclpalettes(files = [tmpfile.name])
@@ -105,6 +116,9 @@ def test_hclpalettes_custom_palconfig_good_custom_div():
             assert isinstance(tmp[k], type(v))
             assert tmp[k] == v
 
+    # Delete temporary file
+    if os.path.exists(tmpfile.name): os.remove(tmpfile.name)
+
 
 # ------------------------------------------
 # We have tested a lot of things in custom_div, so we can
@@ -114,8 +128,11 @@ def test_hclpalettes_custom_palconfig_good_custom_div():
 # ------------------------------------------
 def test_hclpalettes_custom_palconfig_good_custom_seq():
 
-    tmpfile = NamedTemporaryFile()
-    with open(tmpfile.name, "w") as fid: fid.write(good_custom_seq)
+    tmpfile = NamedTemporaryFile(delete = False)
+    with open(tmpfile.name, "w") as fid:
+        fid.write(good_custom_seq)
+        fid.flush()
+        fid.close()
 
     pals = hclpalettes(files = tmpfile.name)
 
@@ -124,6 +141,8 @@ def test_hclpalettes_custom_palconfig_good_custom_seq():
     assert callable(pal.get_settings()["h2"])
     assert pal.get_settings()["h2"](100) == -50.
 
+    # Delete temporary file
+    if os.path.exists(tmpfile.name): os.remove(tmpfile.name)
 
 # ------------------------------------------
 # Reading both at the same time
@@ -131,10 +150,17 @@ def test_hclpalettes_custom_palconfig_good_custom_seq():
 def test_hclpalettes_custom_palconfig_good_custom_seq_and_div():
 
     # Write and read both files at once
-    tmpfile_div = NamedTemporaryFile()
-    with open(tmpfile_div.name, "w") as fid: fid.write(good_custom_div)
-    tmpfile_seq = NamedTemporaryFile()
-    with open(tmpfile_seq.name, "w") as fid: fid.write(good_custom_seq)
+    tmpfile_div = NamedTemporaryFile(delete = False)
+    with open(tmpfile_div.name, "w") as fid:
+        fid.write(good_custom_div)
+        fid.flush()
+        fid.close()
+
+    tmpfile_seq = NamedTemporaryFile(delete = False)
+    with open(tmpfile_seq.name, "w") as fid:
+        fid.write(good_custom_seq)
+        fid.flush()
+        fid.close()
 
     # Reading palette config
[…]
