### 2024-10-01 cc95922
fix: Replacing python setup.py install with pip install .; Updating release version
