### 2024-10-01 45fc530
fix: Adding install setuptools to ensure it is available where needed
