### 2024-10-01 b57788d
fix/devel: Working on a fix to ensure tests on Windows work with both 32/64 bit
