### 2024-10-01 6bac1ee
fix/devel: Working on a fix to ensure tests on Windows work with both 32/64 bit
