### 2024-10-01 31f8efd
fix: Fixing an issue with Windows32 in gtrans/qtrans (colorlib)
