+++ src/colorspace/colorlib.py
@@ -233,12 +233,13 @@ class colorlib:
 
         __fname__ = inspect.stack()[0][3] # Name of this method
 
-        fmax = sys.float_info.max
-
         # Input check
         if isinstance(gamma, float): gamma = np.asarray([gamma])
         if len(gamma) == 1 and not len(gamma) == len(u):
             gamma = np.repeat(gamma, len(u))
+
+        # Convert input to float
+        u = np.asarray(u, dtype = "float")
 
         # Checking inputs
         self._check_input_arrays_(__fname__, u = u, gamma = gamma)
@@ -247,7 +248,7 @@ class colorlib:
         for i,val in np.ndenumerate(u):
             # np.fmax to avoid overflow
             if val > 0.00304:
-                u[i] = np.fmin(fmax, 1.055 * np.power(val, (1. / gamma[i])) - 0.055)
+                u[i] = 1.055 * np.power(val, (1. / gamma[i])) - 0.055
             else:
                 u[i] = 12.92 * val
 
@@ -282,6 +283,9 @@ class colorlib:
 
         # Checking inputs
         self._check_input_arrays_(__fname__, u = u, gamma = gamma)
+
+        # Convert input to float
+        u = np.asarray(u, dtype = "float")
 
         # Transform 
         for i,val in np.ndenumerate(u):
