### 2024-10-01 85f61fc
fix/feat: More clear differenciation of hard and soft tests; hard (mandatory dependencies) and soft (additional/suggested dependencies)
