### 2024-10-01 781a477
fix: Adjusting documentation - installation
