### 2024-10-02 6db7662
fix: 'make clean' now deletes the virtual environments used for testing
