+++ paper.md
@@ -31,9 +31,10 @@ color palettes for qualitative or quantitative (sequential or diverging)
 information. These palettes (as well as any other sets of colors) can be
 visualized, assessed, and manipulated in various ways, e.g., by color swatches,
 emulating the effects of color vision deficiencies, or depicting the perceptual
-properties. Finally, the color palettes generated by the package can be easily
-integrated into standard visualization workflows in Python, e.g., using
-_matplotlib_, _seaborn_, or _plotly_.
+properties. Finally, _colorspace_ integrates seamlessly with standard Python
+visualization libraries like _matplotlib_, _seaborn_, and _plotly_, making it a
+valuable resource for developers and practitioners to customize, assess, and
+implement color palettes in their workflows.
 
 
 # Statement of need
@@ -72,8 +73,10 @@ but also contrast ratios, desaturation, lightening/darkening, etc.
 
 The _colorspace_ Python package was inspired by the eponymous R package
 [@Zeileis:2020]. It comes with extensive documentation at
-<https://retostauffer.github.io/python-colorspace/>, including many practical
-examples. Selected highlights are presented in the following.
+<https://retostauffer.github.io/python-colorspace/>, featuring practical examples
+that complement existing data visualization tools. colorspace enhances the
+toolkit for both casual users and visualization experts. Selected highlights
+are presented in the following.
 
 
 # Key functionality
