### 2024-10-03 5d9a953
fix: Suppressing warnings regarding invalid escape sequences in dostrings (markdown)
