+++ pytest.ini
@@ -6,3 +6,4 @@ markers =
 
 filterwarnings = 
     ignore: FigureCanvasAgg is non-interactive, and thus cannot be shown:UserWarning
+    ignore: invalid escape sequence
