### 2024-10-03 62ec4b8
fix: Adjusting 'Test and development' section (testing strategy/wording)
