### 2024-10-04 fb80724
extend description of target audience and research scope, optimize figure placement
