+++ paper.md
@@ -32,9 +32,9 @@ information. These palettes (as well as any other sets of colors) can be
 visualized, assessed, and manipulated in various ways, e.g., by color swatches,
 emulating the effects of color vision deficiencies, or depicting the perceptual
 properties. Finally, _colorspace_ integrates seamlessly with standard Python
-visualization libraries like _matplotlib_, _seaborn_, and _plotly_, making it a
-valuable resource for developers and practitioners to customize, assess, and
-implement color palettes in their workflows.
+graphics packages like _matplotlib_, _seaborn_, and _plotly_, making it a
+valuable resource for both developers and practitioners to customize, assess, and
+implement color palettes in their data visualization workflows.
 
 
 # Statement of need
@@ -73,10 +73,11 @@ but also contrast ratios, desaturation, lightening/darkening, etc.
 
 The _colorspace_ Python package was inspired by the eponymous R package
 [@Zeileis:2020]. It comes with extensive documentation at
-<https://retostauffer.github.io/python-colorspace/>, featuring practical examples
-that complement existing data visualization tools. colorspace enhances the
-toolkit for both casual users and visualization experts. Selected highlights
-are presented in the following.
+<https://retostauffer.github.io/python-colorspace/>, including many practical
+examples. The package complements existing graphics packages in Python
+both for casual users and data visualization experts. Selected highlights
+are presented in the following, motivating its usefulness for various kinds
+of graphics in different fields of application and research.
 
 
 # Key functionality
@@ -113,6 +114,8 @@ the end of the palette. `pal4` just changes the starting hue
 for the palette to green (200) instead of purple. All four palettes
 are visualized by the `swatchplot` function from the package.
 
+![Swatches of four HCL-based sequential palettes: `pal1` is the predefined HCL-based viridis palette, `pal2` is identical to `pal2` but created "by hand" and `pal3` and `pal4` are modified versions with a triangular chroma paths and reduced hue range, respectively.\label{fig:chosingpalettes}](paper_assets/fig-chosing-palettes.png)
+
 The objects returned by the palette functions provide a series of
 methods, e.g., `pal1.settings` for displaying the HCL parameters,
 `pal1(3)` for obtaining a number of hex colors, or `pal1.cmap()`
@@ -120,13 +123,11 @@ for setting up a _matplotlib_ color map, among others.
 
 ```
 from colorspace import palette, sequential_hcl, swatchplot
-
 pal1 = sequential_hcl(palette = "viridis")
 pal2 = sequential_hcl(h = [300, 75], c = [40, 95], l = [15, 90],
                       power = [1., 1.1])
 pal3 = sequential_hcl(palette = "viridis", cmax = 90,  c2 = 20)
 pal4 = sequential_hcl(palette = "viridis", h1 = 200)
-
 swatchplot({"Viridis (and altered versions of it)": [
                palette(pal1(7), "By name"), 
                palette(pal2(7), "By hand"),
@@ -134,8 +135,6 @@ swatchplot({"Viridis (and altered versions of it)": [
                palette(pal4(7), "With smaller hue range")
            ]}, figsize = (8, 1.75));
 ```
-
-![Swatches of four HCL-based sequential palettes: `pal1` is the predefined HCL-based viridis palette, `pal2` is identical to `pal2` but created "by hand" and `pal3` and `pal4` are modified versions with a triangular chroma paths and reduced hue range, respectively.\label{fig:chosingpalettes}](paper_assets/fig-chosing-palettes.png)
 
[…]
