### 2024-10-06 ec597e0
fix/feat: Providing support for systems without tkinter
