### 2024-10-06 5d90f06
fix: Fixing small issues with demo.Spectrum which showed duplicated axis
