### 2024-10-06 04bad89
feat: Adding additional information for systems where tkinter support is missing.
