### 2024-10-06 6fb1d34
feat: Adding more details to RuntimeError when tkinter stdlib missing
