### 2024-10-06 bcdb0f2
fix: Ignore warnings when matplotlib is running with agg backend
