+++ pytest.ini
@@ -7,3 +7,4 @@ markers =
 filterwarnings = 
     ignore: FigureCanvasAgg is non-interactive, and thus cannot be shown:UserWarning
     ignore: invalid escape sequence
+    ignore: Matplotlib is currently using agg, which is a non-GUI backend, so cannot show the figure
