### 2024-10-06 97bef63
fix: Adding changes to changelog.md (0.5.1)
