### 2024-10-06 d64817d
fix/feat: Updating GUI screenshit, adjusted Tcl/Tk support information
