### 2024-10-06 5da866b
fix: Some additional comments/docstring for situations where tkinter is missing
