### 2024-10-06 e640298
fix: Minor touch-ups on installation guide (testing and development)
