+++ _quarto/installation.qmd
@@ -159,8 +159,8 @@ make venv && make install
 # Activate venv
 source venv/bin/activate
 
-# Running tests: Runs 'hard tests' (minimal requirements) first followed
-# by extended tests (with additional requirements)
+# Running tests: Runs 'hard tests' (minimal requirements) first,
+# followed by extended tests (with additional requirements)
 make test
 
 # Create coverage report; tries to open report in firefox
