### 2024-10-06 234fe62
fix: github workflow for coverage requires pip install -e .
