### 2024-10-07 ef76f1b
Update test_palettes_custom_palconfig_files.py
Small touch-ups on pull request.
