### 2024-10-07 retostauffer/python-colorspace#21 @retostauffer
Dear @hollowscene

thank you for pointing this out and providing a pull request to resolve this issue. We have merged your PR (https://github.com/retostauffer/python-colorspace/pull/22) and successfully ran all tests on both, Windows and Linux.

Some more details can be found here: https://github.com/retostauffer/python-colorspace/pull/22#issuecomment-2396264315
