### 2024-10-07 91bd340
some streamlining of installation and community guidelines
