### 2024-10-07 21f579a
fix: Ensure setuptools is always installed when installing from source
