### 2024-10-07 retostauffer/python-colorspace#23 @retostauffer
Dear @dmreagan,

Thank you for bringing this issue to our attention. We have made several updates to both our documentation and the `Makefile` to ensure that `setuptools` is installed when installing _colorspace_ from source, either on the system or in a virtual environment.

### Documentation updates

* [Installation instructions](https://retostauffer.github.io/python-colorspace/installation.html#installation-instructions): We now recommend running `pip install setuptools` before installing _colorspace_. If `setuptools` is already installed, pip will confirm with a "Requirement already satisfied" message.
* [Testing instructions](https://retostauffer.github.io/python-colorspace/installation.html#testing): Updated to include `setuptools` in the pip install command to ensure it's available for testing.

### Makefile updates

We've added `setuptools` installation steps in the virtual environment setup:

* `hardenv`: Now runs `pip install setuptools` before installing _colorspace_. This environment is used for running minimal requirement tests (`hardtest`; [Makefile#L14](https://github.com/retostauffer/python-colorspace/blob/0083301181f9d21db7473f93e1c9860f91ed4866/Makefile#L14)).
* `venv`: Added `setuptools` to the [`requirements_devel.txt`](https://github.com/retostauffer/python-colorspace/blob/0083301181f9d21db7473f93e1c9860f91ed4866/requirements_devel.txt) file, which is used for extended testing, test coverage, and documentation generation.

These changes should ensure that `setuptools` is installed if it’s not already available.
