### 2024-10-07 retostauffer/python-colorspace#24 @retostauffer
Thank you @dmreagan, for reporting this issue.

We have resolved the problem (see PR https://github.com/retostauffer/python-colorspace/pull/25) and updated the [Testing](https://retostauffer.github.io/python-colorspace/installation.html#testing) instructions accordingly:

* Removed the `source ...` command from the make rule, which caused the "source: not found" error.
* Corrected the order of commands for running tests (see the last line of the [Testing](https://retostauffer.github.io/python-colorspace/installation.html#testing) section).
