+++ _quarto/articles/choose_palette.qmd
@@ -15,10 +15,6 @@ choose_palette()
 
 ![](/images/img_gui.jpeg){fig-align="center" .img-padding-tb}
 
-If `tkinter` is missing, a `RuntimeError` will be thrown
-(see [Installation: Tcl/Tk support](../installation.qmd#tkinter-support)
-for details).
-
 
 Web App
 -------
