### 2024-10-07 ba6e069
fix: Missing import in choose_palette() when matplotlib is not installed
