### 2024-10-07 4c89e74
fix: Release 0.5.2
