+++ src/colorspace/version.py
@@ -1,3 +1,3 @@
-version = '0.5.1'
-short_version = '0.5.1'
+version = '0.5.2'
+short_version = '0.5.2'
 isreleased = False
