### 2024-10-07 #7120 @retostauffer
Dear @jromanowska, @hollowscene, and @dmreagan,

First of all, we would like to sincerely thank you for handling our submission, as well as for the constructive and valuable feedback on our work. We apologize for the delayed response due to vacations and covid.

We have thoroughly addressed all comments from the review as well as the [pre-review](https://github.com/openjournals/joss-reviews/issues/7053) (handled by Daniel S. Katz).

We have also updated the README to acknowledge the contributions of the reviewers and editors involved in this process. Your efforts in enhancing the quality of our work are greatly appreciated.

## Paper

* Highlight the target audience as well as the research scope of the software. Specifically, we have updated the end of the "Summary" (see [paper.md#L34-L37](https://github.com/retostauffer/python-colorspace/blob/v0.5.2/paper.md?plain=1#L34-L37)) and the last paragraph of the 'Statement of need' ([paper.md#L77-L80](https://github.com/retostauffer/python-colorspace/blob/v0.5.2/paper.md?plain=1#L77-L80)).
* Minor touch-ups for layouting purposes.

## Software

### Key Changes

* New release (`v0.5.2`) available on [PyPI](https://pypi.org/project/colorspace/).
* Updated the [Installation](https://retostauffer.github.io/python-colorspace/installation.html) guidelines to address two issues reported by @dmreagan (see [Issue 23](https://github.com/retostauffer/python-colorspace/issues/23) and [Issue 24](https://github.com/retostauffer/python-colorspace/issues/24)), with the corresponding pull request by @zeileis (see [PR 25](https://github.com/retostauffer/python-colorspace/pull/25)).
* Fixed an issue where tests using temporary files failed on Windows, as reported by @hollowscene (see [Issue 21](https://github.com/retostauffer/python-colorspace/issues/21), addressed in [PR 22](https://github.com/retostauffer/python-colorspace/pull/22)).
* Addressed the issue that installation from source, as described in the documentation, failed on systems where `setuptools` was not installed (reported by @dmreagan in [Issue 23](https://github.com/retostauffer/python-colorspace/issues/23)):
    - Updated the [documentation](https://retostauffer.github.io/python-colorspace/installation.html) to ensure `setuptools` is installed when needed.
    - Included `setuptools` in the `Makefile` and `requirements_devel.txt` to ensure it is available when installing _colorspace_ from source.
    - Added a check to verify that `setuptools` is installed when running `setup.py`.

### Additional Changes

- Fixed a few bugs affecting 32-bit systems in `statshelper.py` and `colorlib.py`.
- Adjusted _colorspace_ to function correctly even when the standard library `tkinter` is missing.
- Fixed a minor visualization bug in `demos.Spectrum`.
- Replaced the deprecated `python setup.py install` command with `pip install .`.
- Streamlined the [Installation](https://retostauffer.github.io/python-colorspace/installation.html)
  and [Community guidelines](https://retostauffer.github.io/python-colorspace/community.html) sections in the documentation.
- Added dedicated rules to run tests for hard and soft dependencies (`make hardtest`, `make softtest`).

Once again, thank you for your time and thoughtful feedback. We look forward to your further comments and suggestions.

Best regards,
@retostauffer and @zeileis
