### 2024-10-21 #7120 @retostauffer
Thanks, @jromanowska!

@hollowscene & @dmreagan, if you want to, you can respond to the individual issues that you raised and we have answered (https://github.com/retostauffer/python-colorspace/issues/21, https://github.com/retostauffer/python-colorspace/issues/23, https://github.com/retostauffer/python-colorspace/issues/24). Feel free to close them if you feel that they have been sufficiently addressed. Or if you respond here, we can also close them in our repository.

Thanks again for your feedback and consideration.
