### 2024-10-22 retostauffer/python-colorspace#23 @dmreagan
Looks good to me. Closing.
