### 2024-10-22 retostauffer/python-colorspace#24 @dmreagan
Looks good to me. Closing.
