### 2024-10-25 0fc81d0
fix: Installing scipy as it is required by plotly create_distplot
