### 2024-10-28 9775d3b
Major release 1.0.0: Updated affiliation metadata and date in the JOSS article, added short note in Changelog.
