### 2024-10-28 retostauffer/python-colorspace#21 @retostauffer
Many thanks again @hollowscene for [reviewing our submission](https://github.com/openjournals/joss-reviews/issues/7120) to [The Open Source Journal](https://https://joss.theoj.org/).

Your bug report and PR (#22) were greatly appreciated! I’ll be closing the issue at this time.
