### 2024-10-28 4e2094d
Major release 1.0.0: Releasing first major version of the Python colorspace package
