+++ src/colorspace/version.py
@@ -1,3 +1,3 @@
 version = '1.0.0'
 short_version = '1.0.0'
-isreleased = False
+isreleased = True
