### 2024-10-28 7496a78
fix: Adding CITATION.cff for zenodo
