### 2024-10-28 9722dbc
dev: Downgrading CITATION.cff to 1.1.0
