### 2024-10-28 fde2092
dev: Going with the very basic CITATION.cff suggested by zenodo themselves
