### 2024-10-28 a308c4b
fix: Given zenodo does not comply with what I tried, although the cff is a valid 1.2.0 cff, reverting back, using zip release upload.
