### 2024-10-28 ac86b6f
fix: Fixing CITATION.cff again, putting license into a string rather than an array
