### 2024-10-28 cbfc37b
Create CITATION.cff
Well ... I need to crate the release via zenodo, not on GitHub. Adding CITATION.cff (v1.2.0) and see if it now accepts my metadata.
