### 2024-10-28 #7120 @retostauffer
- [x] Added 'Austria' to our affiliations, as noted in [@jromanowska's comment](https://github.com/openjournals/joss-reviews/issues/7120#issuecomment-2440834105).
- [x] Verified that all authors' names, affiliations, and ORCIDs are correct in the paper.
- [x] Released version `1.0.0` for the JOSS publication.
  - Minor updates include changes to the changelog and version release (see [git diff](https://github.com/retostauffer/python-colorspace/compare/v0.5.2...v1.0.0)).
- [x] Published new release `1.0.0` on [PyPI](https://pypi.org/project/colorspace) and [Zenodo](https://zenodo.org/records/14004295):
  - **Zenodo/PyPI Version:** `1.0.0`
  - **Zenodo DOI:** [10.5281/zenodo.14004295](https://doi.org/10.5281/zenodo.14004295)
- [x] Confirmed that Zenodo's title, author list, ORCIDs, and affiliations match those in the JOSS article.
- [x] Confirmed that Zenodo's license is consistent with the software license.

To clarify: The `editorialbot` flagged a missing DOI ([10.1016/0306-4573(91)90096-5](https://doi.org/10.1016/0306-4573(91)90096-5)), which actually references a review of the book, not the book itself (see [further discussion](https://github.com/openjournals/joss-reviews/issues/7053#issuecomment-2258039140)).

Finally, we (and I believe I can also speak for @zeileis here) would like to express our gratitude to everyone involved for your time and efforts, from pre-screening to reviewing and managing our submission. This is our first submission to JOSS, as well as our first git-based submission, and it certainly won't be our last!

If we’ve overlooked anything, please let us know. Thank you again and all the best!
@zeileis and @retostauffer
