### 2024-10-29 #7120 @zeileis
Quick feedback regarding the DOI suggested as MISSING above. This was already addressed in the pre-review (see https://github.com/openjournals/joss-reviews/issues/7053#issuecomment-2258039140). The DOI should not be included as it is for a review of the book and not the book itself.

I also looked at the PDF and XML files and just noticed a minor aspect: https://github.com/openjournals/joss-papers/pull/6062#pullrequestreview-2401480863
