### 2024-10-29 #7120 @zeileis
Thank you so much @danielskatz, very much appreciated. We also tried to incorporate your comments from the pre-review https://github.com/openjournals/joss-reviews/issues/7053#issuecomment-2276373641 in the paper. Happy to make further improvements.

Regarding the copyright year: It's not a big issue. I just thought it was worth mentioning. Reto told me (off-list) that it might be due to the copyright 2022-2024 in the `setup.py` of the package (https://github.com/retostauffer/python-colorspace/blob/main/setup.py#L4-L5)
