### 2024-10-29 #7120 @danielskatz
@zeileis - I'm suggesting some minor changes for language in https://github.com/retostauffer/python-colorspace/pull/26.

Additionally, I wonder about the lack of an acknowledgements section. Is there any funding/support or anything else that should be acknowledged?
