### 2024-10-29 0d7e419
fix/feat: Updating ORCID iD style in README.md, adding citations (paper and zenodo archive).
