### 2024-10-29 ff3aa02
Completing ORCID iDs in README.md and _quart/index.qmd
