+++ README.md
@@ -39,40 +39,43 @@ The package provides a Python 3 version of the
 ## Contributors
 
 - [Reto Stauffer](https://retostauffer.org/)
-  <a target="_new" href="https://orcid.org/0000-0002-3798-5507"><img src="https://raw.githubusercontent.com/retostauffer/python-colorspace/main/_quarto/logo-orcid.svg" width="18px" style="position: relative; top: 3px;" alt="ORCID iD"/></a>,
+  <a target="_new" href="https://orcid.org/0000-0002-3798-5507"><img src="https://raw.githubusercontent.com/retostauffer/python-colorspace/main/_quarto/logo-orcid.svg" width="16px" alt="ORCID iD"/></a>,
   (_principal author & maintainer_), Universität Innsbruck, Austria
 - [Achim Zeileis](https://www.zeileis.org/)
-  <a target="_new" href="https://orcid.org/0000-0003-0918-3766"><img src="https://raw.githubusercontent.com/retostauffer/python-colorspace/main/_quarto/logo-orcid.svg" width="18px" style="position: relative; top: 3px;" alt="ORCID iD"/></a>,
+  <a target="_new" href="https://orcid.org/0000-0003-0918-3766"><img src="https://raw.githubusercontent.com/retostauffer/python-colorspace/main/_quarto/logo-orcid.svg" width="16px" alt="ORCID iD"/></a>,
   Universität Innsbruck, Austria
 
 We would like to thank the following individuals for their contributions to the
 review process for this software, as part of the submission to
 [The Journal of Open Source Software](https://joss.theoj.org/) (JOSS):
-[Julia Romanowska @jromanowska, Editor](https://github.com/jromanowska),
-[Daniel S. Katz @danielskatz, Editor](https://github.com/danielskatz),
-[Andrew Ting @hollowscene, Reviewer](https://github.com/hollowscene), and
-[David Reagan @dmreagan, Reviewer](https://github.com/dmreagan).
+[Julia Romanowska @jromanowska, Editor](https://github.com/jromanowska)
+  <a target="_new" href="https://orcid.org/0000-0001-6733-1953"><img src="https://raw.githubusercontent.com/retostauffer/python-colorspace/main/_quarto/logo-orcid.svg" width="16px" alt="ORCID iD"/></a>,
+[Daniel S. Katz @danielskatz, Editor](https://github.com/danielskatz)
+  <a target="_new" href="https://orcid.org/0000-0001-5934-7525"><img src="https://raw.githubusercontent.com/retostauffer/python-colorspace/main/_quarto/logo-orcid.svg" width="16px" alt="ORCID iD"/></a>,
+[David Reagan @dmreagan, Reviewer](https://github.com/dmreagan)
+  <a target="_new" href="https://orcid.org/0000-0002-8359-7580"><img src="https://raw.githubusercontent.com/retostauffer/python-colorspace/main/_quarto/logo-orcid.svg" width="16px" alt="ORCID iD"/></a>,
+[Andrew Ting @hollowscene, Reviewer](https://github.com/hollowscene).
 
 Furthermore, the Python package implements the functionality from the R package
 that is co-written with
 [Ross Ihaka](https://www.stat.auckland.ac.nz/~ihaka/),
 [Paul Murrell](https://www.stat.auckland.ac.nz/~paul/)
[…]
