### 2024-10-29 #7120 @retostauffer
@danielskatz, thank you for proofreading and for your suggested changes in PR https://github.com/retostauffer/python-colorspace/pull/26, now merged. In addition to these changes, we also slightly cropped Figure 4 for layout purposes.

*Acknowledgements:* This work was conducted independently, without support from any funding agency or research project. We have no additional acknowledgments beyond the editorial and review process, but we can add an acknowledgment for the editors and reviewers if desired.

*Minor changes:* Added citations (and ORCID iDs) to the [README.md](https://github.com/retostauffer/python-colorspace/blob/main/README.md) as well as the [documentation landing page](https://github.com/retostauffer/python-colorspace/blob/main/_quarto/index.qmd).

If there is anything more to do, please let us know. All the best,
@zeileis and @retostauffer
