### 2024-10-29 #7120 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
