### 2024-10-29 #7120 @danielskatz
Congratulations to @retostauffer (Reto Stauffer) and @zeileis on your publication!!
