### 2024-12-20 #7120 @zeileis
Daniel @danielskatz & Julia @jromanowska, I had tried to make a donation via the link above when our paper was published and I now tried again. However, I cannot successfully load the page, it keeps telling me:

> (If this page appears blank, please disable your ad/popup-blocker to view our donation form. Thanks!)

I tried Firefox and Chromium on Debian as well as Samsung Internet and Firefox on Android, enabling ads, pop-ups, cookies, etc. but nothing works. By now I'm afraid that the time spent on this exceeds the worth of the donation I intended to make (out of my personal pocket).

Maybe it is possible to make donations a little bit simpler or more accessible in the future? Then I'm happy to contribute.

In any case, let me thank you again. I very much appreciate the service you provide to the science in general and the open-source software community in particular.

Season's Greetings, Achim
