### 2024-12-20 #7120 @danielskatz
👋 @zeileis - https://numfocus.org/donate-to-open-journals should work in the meantime - (Open Journals is the official org the develops editorialbot and runs JOSS, JOSE, etc. and is the subunit under NumFOCUS that has the governance for JOSS)  And thanks for pointing this out - we'll get it fixed.
