### 2024-12-20 #7120 @danielskatz
@zeileis - the original donate link should now be working again.  Thanks again for pointing this out so we could get it fixed.
