### 2024-12-20 #7120 @retostauffer
@danielskatz, thank you very much for your reply!

I can confirm that the link as well as the donation process is working flawlessly (Linux, Chromium). I found some extra funds in my pocket to support JOSS (anonymously, to keep things clean).

Both of us are passionate supporters of excellent open-source software made available to everyone at no charge. We truly appreciate what you do and will continue to do in the future.

I may be repeating myself, but it was a great and very smooth experience working with you (and JOSS)!
