### 2024-12-20 #7120 @zeileis
:100:

I also just donated, the link works fine now. Thanks for taking care of it so quickly. Maybe you want to let your sister journals know as well. https://numfocus.org/donate-to-jose still seems to have the same problem.
